+++ airplane/tests/test_airplane.py
@@ -1,14 +1,16 @@
 import os, shutil
 
 from django.conf import settings
+from django.core.management import call_command
 from django.template import Context, Template
 from django.test import TestCase, override_settings
 import six
 
 import airplane
-from airplane.templatetags.airplanetags import _convert_url
+from airplane.utils import convert_url
 from mock import patch
-from wrench.contexts import temp_directory
+from context_temp import temp_directory
+from waelstow import capture_stdout
 
 # ============================================================================
 # Test Objects
@@ -42,6 +44,9 @@ class AirplaneTests(TestCase):
     def tearDown(self):
         self._remove_local_caches()
 
+    def assertFile(self, dir_name, filename):
+        self.assertTrue(os.path.exists(os.path.join(dir_name, filename)))
+
     def _remove_local_caches(self):
         if os.path.exists(self.cache1_path):
             shutil.rmtree(self.cache1_path)
@@ -58,20 +63,22 @@ class AirplaneTests(TestCase):
         result = template.render(context)
         self.assertEqual(expected, result.strip())
 
-    def _check_build_cache(self, url, expected, dir_name):
+    def _check_build_cache(self, url, expected_url, dir_name, 
+            expected_filename):
         # test with a mocked request that works
         with patch('requests.get') as mock_requests:
             mock_requests.side_effect = fake_get
 
             # render the tag
-            self._render(url, expected)
+            self._render(url, expected_url)
 
             # check that the directory got a file
-            os.path.exists(os.path.join(dir_name, expected))
+            self.assertFile(dir_name, expected_filename)
 
     def test_airplane(self):
         url = 'http://foo.com'
-        expected = '/static/' + _convert_url(url)
+        expected_filename = convert_url(url)
+        expected_url = '/static/' + expected_filename
 
         # test pass through, DEBUG=False, no AIRPLANE_MODE set
         self._render(url, url)
@@ -88,7 +95,8 @@ class AirplaneTests(TestCase):
                     AIRPLANE_MODE=airplane.BUILD_CACHE):
 
                 # test with a mocked request that works
-                self._check_build_cache(url, expected, td)
+                self._check_build_cache(url, expected_url, td, 
+                    expected_filename)
 
                 # test with a mocked request that fails (404s etc)
                 with patch('requests.get') as mock_requests:
@@ -96,7 +104,7 @@ class AirplaneTests(TestCase):
 
                     # render the tag
                     with self.assertRaises(IOError):
-                        self._render(url, expected)
+                        self._render(url, expected_url)
 
         # test cache creation with a local directory that needs creating
         self._remove_local_caches()
@@ -106,10 +114,8 @@ class AirplaneTests(TestCase):
                 DEBUG=True,
                 AIRPLANE_MODE=airplane.BUILD_CACHE):
 
-            self._check_build_cache(url, expected, self.cache1)
-
-            # check that the directory got a file
-            os.path.exists(os.path.join(self.cache1, expected))
+            self._check_build_cache(url, expected_url, self.cache1_path, 
+                expected_filename)
 
         # test with a non-default cache name that is local
         with override_settings(
@@ -117,7 +123,34 @@ class AirplaneTests(TestCase):
                 AIRPLANE_CACHE=self.cache2,
                 AIRPLANE_MODE=airplane.BUILD_CACHE):
 
-            self._check_build_cache(url, expected, self.cache2)
+            self._check_build_cache(url, expected_url, self.cache2_path, 
+                expected_filename)
 
-            # check that the directory got a file
-            os.path.exists(os.path.join(self.cache1, expected))
+    def test_commands(self):
+        # test aircache command
+        with temp_directory() as td:
+            with override_settings(
+                    DEBUG=True,
+                    AIRPLANE_CACHE=td):
+
+                with patch('requests.get') as mock_requests:
+                    mock_requests.side_effect = fake_get
+
+                    url = 'http://foo.com/thing'
+                    call_command('aircache', url)
+
+                    filename = convert_url(url)
+
+                    # check that the directory got a file
+                    self.assertTrue(os.path.exists(os.path.join(td, filename)))
+
+                # run airinfo and make sure it doesn't blow up
+                with capture_stdout():
+                    with override_settings(AIRPLANE_MODE=airplane.BUILD_CACHE):
+                        call_command('airinfo')
+
+                    with override_settings(AIRPLANE_MODE=airplane.USE_CACHE):
+                        call_command('airinfo')
+
+                    with override_settings(AIRPLANE_MODE=0):
+                        call_command('airinfo')