+++ CHANGES.rst
@@ -1,3 +1,10 @@
+0.2
+===
+
+* **BREAKING CHANGE**: changed cache filename format
+* added management commands for listing and adding to cache
+* changed testing to be against django 1.11, 2.0 
+
 0.1.1
 =====